requirements.txt
# Django
# Django==1.8.7
six==1.10.0

##### THESE ARE VERSION LOCKED PACKAGES FROM
# https://github.com/GeoNode/geonode/blob/2.6.x/requirements.txt
#    at commit https://github.com/GeoNode/geonode/commit/b1dcd7c4974d6e6cb642b7e9a54f7b7370eb4758
Django==1.8.18
MultipartPostHandler==0.1.0
OWSLib==0.11.0
Paver==1.2.4
Unidecode==0.4.19
amqp==1.4.9
anyjson==0.3.3
awesome-slugify==1.6.2
beautifulsoup4==4.2.1
billiard==3.3.0.23
boto==2.38.0
celery==3.1.18
dj-database-url==0.4.2
django-activity-stream==0.6.1
django-appconf==0.5
django-autocomplete-light==2.3.3
django-bootstrap3-datetimepicker==2.2.3
django-braces==1.10.0
django-celery==3.1.16
django-downloadview==1.2
django-extensions==1.6.1
django-forms-bootstrap==3.0.1
django-friendly-tag-loader==1.2.1
django-geoexplorer==4.0.5
django-guardian==1.4.6
django-haystack==2.4.1
django-jsonfield==0.9.16
django-jsonfield-compat==0.4.4
django-leaflet==0.13.7
django-modeltranslation==0.12
django-oauth-toolkit<1.0
django-mptt==0.8.6
django-nose==1.4.4
django-pagination==1.0.7
django-polymorphic==0.9.2
django-storages==1.1.8
django-taggit==0.21.0
django-tastypie==0.12.2
flake8==2.3.0
geolinks==0.2.0
geonode-agon-ratings==0.3.5
geonode-announcements==1.0.8
geonode-arcrest==10.2
geonode-avatar==2.1.6
geonode-dialogos==0.7
geonode-notification==1.1.1
geonode-user-accounts==1.0.15
geonode-user-messages==0.1.5
gisdata==0.5.4
gsconfig==1.0.6
gsimporter==1.0.0
httplib2==0.10.3
kombu==3.0.35
lxml==3.6.4
mccabe==0.5.2
oauthlib==2.0.1
pep8==1.6.2
pillow==3.3.1
pinax-theme-bootstrap==3.0a11
pinax-theme-bootstrap-account==1.0b2
pycsw==2.0.2
elasticsearch==2.4.0
pyflakes==1.2.3
pyproj==1.9.5.1
python-dateutil==2.5.3
python-mimeparse==1.5.2
pytz==2016.6.1
pyyaml==3.11
regex==2016.7.21
requests==2.11.1
transifex-client==0.10
xmltodict==0.9.2
django-treebeard==3.0
django-geonode-client==0.0.15


# Geonode
# -e git+https://github.com/NJCoast/geonode.git@2.6.x#egg=geonode
# geonode==2.6.3
-e git+https://github.com/GeoNode/geonode.git@2.6.3#egg=geonode

# Gdal
# pygdal==1.10.1

# Configuration
dj-database-url==0.4.2
python-decouple==3.0
whitenoise==3.3.0

# database
psycopg2

# Annotations and sockets
channels>=1.1,<2
